+++ packages/core/src/data_structures/index.ts
@@ -1,1 +1,2 @@
+export { default as SelectorSet } from './selector_set';
 export { default as SetMap } from './set_map';