+++ packages/core/src/index.ts
@@ -6,7 +6,3 @@ export * from './element';
 export * from './events';
 export * from './lazy_import';
 export * from './lifecycle';
-export * from './observers/attribute_observer';
-export * from './observers/element_observer';
-export * from './observers/selector_observer';
-export * from './observers/token_list_observer';